+++ src/EphIt/Classlibraries/EphIt.Db.Models/EphIt.Db.Models.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\EphIt.Db\EphIt.Db.csproj" />
+    <PackageReference Include="System.ComponentModel.Annotations" Version="4.7.0" />
   </ItemGroup>
 
 </Project>
