+++ src/EphIt/Classlibraries/EphIt.Db.Models/Job.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
@@ -35,22 +33,5 @@ namespace EphIt.Db.Models
         public virtual JobParameters JobParameters { get; set; }
         public virtual ICollection<JobOutput> JobOutput { get; set; }
     }
-    public class JobConfiguration : IEntityTypeConfiguration<Job>
-    {
-        public void Configure(EntityTypeBuilder<Job> builder)
-        {
-            builder.HasOne(d => d.CreatedByUser)
-                .WithMany(many => many.Job)
-                .OnDelete(DeleteBehavior.Restrict);
-            builder.HasOne(d => d.JobStatus)
-                .WithMany(p => p.Job)
-                .HasForeignKey(key => key.JobStatusId)
-                .OnDelete(DeleteBehavior.Restrict);
-            builder.HasOne(d => d.ScriptVersion)
-                .WithMany(p => p.Jobs)
-                .HasForeignKey(key => key.ScriptVersionId)
-                .OnDelete(DeleteBehavior.Restrict);
-           
-        }
-    }
+    
 }
