+++ src/EphIt/Classlibraries/EphIt.Db.Models/JobOutput.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
@@ -18,14 +16,5 @@ namespace EphIt.Db.Models
 
         public virtual Job Job { get; set; }
     }
-    public class JobOutputConfiguration : IEntityTypeConfiguration<JobOutput>
-    {
-        public void Configure(EntityTypeBuilder<JobOutput> builder)
-        {
-            builder.HasOne(p => p.Job)
-                .WithMany(m => m.JobOutput)
-                .HasForeignKey(key => key.JobUid)
-                .OnDelete(DeleteBehavior.Cascade);
-        }
-    }
+    
 }
