+++ src/EphIt/Classlibraries/EphIt.Db.Models/RoleMembershipUser.cs
@@ -1,0 +1,18 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace EphIt.Db.Models
+{
+    public class RoleMembershipUser
+    {
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int RoleMembershipUserId { get; set; }
+        public int RoleId { get; set; }
+        public int UserId { get; set; }
+
+        public virtual Role Role { get; set; }
+        public virtual User User { get; set; }
+    }
+    
+}
