+++ src/EphIt/Classlibraries/EphIt.Db/EphIt.Db.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.7.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\EphIt.Db.Models\EphIt.Db.Models.csproj" />
+  </ItemGroup>
+
 </Project>
